Add stop logic to run method
The stop logic should be under user control. By putting it in a
method its more easily accessible, instead of changing the template
class in the example folder.

The downside is that the EndlessViewIterator goes into a while loop
whichs needs to be broken. As the logic in the run method is one
level higher, its passed into the EVI if its being used. The while
loop can get broken and stop the iterator after the stop condition is
reached.

The flow goes as follows:
EVI.__next__ is called when the iterator is empty
try: EVI.next calls the ViewIterator.next
which calls TaskViewIterator.__next__
which calls TaskViewIterator.claim_task
which calls _claim_task which fails and
throws a StopIteration at EVI.next
and the while-loop continues.