Update processing to be more general
The eval implementation assumes certain forms for input
which results in complex strings from the token to be
parsed into commands. This is unwanted, as programs can
generate quite complicated commands in string form, with
different shapes (for safety for ex.) like single-, double-
quotes, backticks, and other characters that bash, python
and json may interpret/parse differently.

The implementation in this commit changes the bash part to
leave the input string intact, and not split it into pieces.
Moreover, the shell part of python is changed into default
behaviour of subprocess, along with listing the input,
which also helps with parsing complex strings properly.

This was tested with string containing escaped characters
in nested and nonnested forms, and is more robust.